+++ erp-model/erp-model-mrp/src/main/java/com/erp/model/mrp/vo/RptOutOfStockVO.java
@@ -26,7 +26,7 @@ public class RptOutOfStockVO {
     /**
      * 销量
      */
-    private Integer salesQty;
+    private BigDecimal salesQty;
 
     /**
      * 金额
