+++ erp-model/erp-model-mrp/src/main/java/com/erp/model/mrp/vo/SalesAnalysisVO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -38,6 +39,6 @@ public class SalesAnalysisVO {
         /**
          * 数量
          */
-        private List<Integer> qty;
+        private List<BigDecimal> qty;
     }
 }
